requirements.txt
numpy==1.26.4
sympy>=1.12
python-docx==1.1.0 